backend/requirements.txt
fastapi
uvicorn
huggingface-hub
InstructorEmbedding
faiss-cpu
llama-index>=0.11.0
llama-index-embeddings-huggingface
llama-index-llms-llama-cpp
llama-index-vector-stores-faiss
pdfplumber
pydantic
pymongo
requests
python-docx
python-multipart
llama-index-embeddings-fastembed
beautifulsoup4
streamlit
llama-cpp-python==0.2.54
crewai
langchain-community
playwright
urllib3
asyncio
